Move RewardType out of runtime
